Сделала каталог пиццы в отдельном фрагменте
