+++ app/src/main/res/layout/activity_main.xml
@@ -9,48 +9,8 @@
     android:orientation="vertical">
 
     <FrameLayout
-        android:id="@+id/frameInfoPizza"
+        android:id="@+id/framePizza"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="100dp"
-        android:orientation="horizontal">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            style="@style/textStyleTitle"
-            android:text="@string/Pizza"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            android:layout_marginStart="30dp"
-            />
-
-        <Button
-            android:id="@+id/basket"
-            android:layout_width="40dp"
-            android:layout_height="40dp"
-            android:background="@drawable/baseline_shopping_basket_50"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            android:layout_marginEnd="30dp"
-
-            />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <ListView
-        android:id="@+id/allPizza"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="30dp"
-        tools:listitem="@layout/list_item"
-        android:layout_marginHorizontal="10dp"
-        android:layout_marginBottom="30dp"
-        >
-    </ListView>
+        android:layout_height="match_parent"/>
 
 </LinearLayout>
