+++ app/src/main/res/layout/fragment_pizza_details.xml
@@ -7,7 +7,7 @@
 
     <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="800dp"
+        android:layout_height="match_parent"
         android:orientation="vertical"
         android:layout_marginHorizontal="10dp">
